cloud-function-trigger/requirements.txt
google-cloud-run==0.10.0
google-cloud-pubsub==2.18.0
